requirements.txt
playwright
beautifulsoup4
langchain
langchain-openai
chromadb
pandas
python-dotenv
pydantic
html2text
streamlit
